+++ analytical/templatetags/analytical.py
@@ -27,6 +27,7 @@ TAG_MODULES = [
     'analytical.optimizely',
     'analytical.performable',
     'analytical.reinvigorate',
+    'analytical.woopra',
 ]